chore: rename lib/gui/constants.h => lib/gui/Constants.h
