+++ src/lib/gui/messages.cpp
@@ -6,9 +6,9 @@
 
 #include "messages.h"
 
+#include "Constants.h"
 #include "Logger.h"
 #include "common/constants.h"
-#include "constants.h"
 #include "env_vars.h"
 #include "styles.h"
